feat(alerts): demo-suppression flag for advanced controls (#407)
* feat(alerts): gate Auto-fire + Subscribe behind NEXT_PUBLIC_ALERTS_ADVANCED for demo

* fix(tests): update AlertsPage test import after component extraction

* ci: retrigger check after test import fix

* fix(test): update empty state assertion to match new component copy

---------
