+++ .env.example
@@ -52,3 +52,7 @@ VERCEL_PROTECTION_BYPASS=
 TEST_USER_EMAIL=e2e@atlas.test
 TEST_USER_PASSWORD=e2etest123
 PLAYWRIGHT_BASE_URL=http://localhost:3000
+
+# Alerts advanced controls (Auto-fire + Subscribe)
+# Flip to "true" to restore advanced buttons for full release.
+NEXT_PUBLIC_ALERTS_ADVANCED=false
